requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
databases>=0.5.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
openai>=1.3.7
httpx>=0.24.0
pydantic>=2.0.0
python-jose==3.3.0
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.13.1
tabulate==0.9.0 